Refactor to separate comment logic and line graph
